Version 2.0
This version now uses Cobra for the command line interface. The code has been cleaned up and the pkg/ directory is now an API that could be used by other projects. The command line tool now has two modes update and updater. The Dockerfile is now non-root and Go Dep is being used for vendoring.